+++ helios/drivers/serial.c
@@ -1,5 +1,27 @@
-#include "../arch/x86_64/ports.h"
+/**
+ * @file drivers/serial.c
+ *
+ * Copyright (C) 2025  Dylan Parks
+ *
+ * This file is part of HeliOS
+ *
+ * HeliOS is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 #include <drivers/serial.h>
+
+#include "../arch/x86_64/ports.h"
 
 /**
  * @brief Initializes the serial port for communication.
@@ -14,28 +36,31 @@
  */
 int init_serial(void)
 {
-    outb(PORT + 1, 0x00); // Disable all interrupts
-    outb(PORT + 3, 0x80); // Enable DLAB (set baud rate divisor)
-    outb(PORT + 0, 0x03); // Set divisor to 3 (lo byte) 38400 baud
-    outb(PORT + 1, 0x00); //                  (hi byte)
-    outb(PORT + 3, 0x03); // 8 bits, no parity, one stop bit
-    outb(PORT + 2, 0xC7); // Enable FIFO, clear them, with 14-byte threshold
-    outb(PORT + 4, 0x0B); // IRQs enabled, RTS/DSR set
-    outb(PORT + 4, 0x1E); // Set in loopback mode, test the serial chip
-    outb(PORT + 0, 0xAE); // Test serial chip (send byte 0xAE and check if serial returns same byte)
+	outb(PORT + 1, 0x00); // Disable all interrupts
+	outb(PORT + 3, 0x80); // Enable DLAB (set baud rate divisor)
+	outb(PORT + 0, 0x03); // Set divisor to 3 (lo byte) 38400 baud
+	outb(PORT + 1, 0x00); //                  (hi byte)
+	outb(PORT + 3, 0x03); // 8 bits, no parity, one stop bit
+	outb(PORT + 2, 0xC7); // Enable FIFO, clear them, with 14-byte threshold
+	outb(PORT + 4, 0x0B); // IRQs enabled, RTS/DSR set
+	outb(PORT + 4, 0x1E); // Set in loopback mode, test the serial chip
+	outb(PORT + 0, 0xAE); // Test serial chip (send byte 0xAE and check if serial returns same byte)
 
-    // Check if serial is faulty (i.e: not same byte as sent)
-    if (inb(PORT + 0) != 0xAE) {
-        return 1;
-    }
+	// Check if serial is faulty (i.e: not same byte as sent)
+	if (inb(PORT + 0) != 0xAE) {
+		return 1;
+	}
 
-    // If serial is not faulty set it in normal operation mode
-    // (not-loopback with IRQs enabled and OUT#1 and OUT#2 bits enabled)
-    outb(PORT + 4, 0x0F);
-    return 0;
+	// If serial is not faulty set it in normal operation mode
+	// (not-loopback with IRQs enabled and OUT#1 and OUT#2 bits enabled)
+	outb(PORT + 4, 0x0F);
+	return 0;
 }
 
-static int is_transmit_empty(void) { return inb(PORT + 5) & 0x20; }
+static int is_transmit_empty(void)
+{
+	return inb(PORT + 5) & 0x20;
+}
 
 /**
  * @brief Writes a character to the serial port.
@@ -47,9 +72,9 @@ static int is_transmit_empty(void) { return inb(PORT + 5) & 0x20; }
  */
 void write_serial(char a)
 {
-    while (is_transmit_empty() == 0)
-        ;
-    outb(PORT, a);
+	while (is_transmit_empty() == 0)
+		;
+	outb(PORT, (uint8_t)a);
 }
 
 /**
@@ -62,6 +87,6 @@ void write_serial(char a)
  */
 void write_serial_string(const char* s)
 {
-    while (*s)
-        write_serial(*s++);
+	while (*s)
+		write_serial(*s++);
 }
